✨: 마이페이지 depth 추가
